New indents broke rc.local md5sum, startup script
The indentation was introduced in cad2e0e11a9321cec88534a5537fe3157f522f0b, not when the safety hash check was added in 2c6f069c2f5092616a9e60677d6ea056c8b292a6, my first guess.
The indentations makes the rc.local begin with spaces instead of `#!/bin/bash` and will not run. But it gets better.
The md5 safety check is based on rc.local without indents, so after the added indentations the if-block will never be skipped. If the script is run a second time, the original rc.local backed up in rc.local.bak will be overwritten, and on boot result in a never ending loop, trying to expand the filesystem each time. A user could run the pishrink.sh a second time on a disk image, e.g. after making more edits.