reverted to using raspi-config to be used for automatic image resizing
while not having to rely on the raspi-config tool does make the pishrink utility more portable, the raspi-config tool is already properly maintained. rather than being forced to maintain the autosizing utility in this script (which has already been broken for several months due to a linux kernal change), it just makes more sense to not have to rewrite properly maintained code.